+++ rossum-mcp/CHANGELOG.md
@@ -11,4 +11,5 @@ All notable changes to this project will be documented in this file.
 - Added `get_hook` method for `rossum_api.models.hook.Hook` objects. [#49](https://github.com/stancld/rossum-mcp/pull/49)
 - Added `get_rule` method for `rossum_api.models.rule.Rule` objects. [#49](https://github.com/stancld/rossum-mcp/pull/49)
 - Added `get_engine_fields` method for `rossum_api.models.engine.EngineFields` objects. [#49](https://github.com/stancld/rossum-mcp/pull/49)
-- Added `get_relation` and `list_relations` methods for `rossum_api.models.relation.Relations` objects. [#50](https://github.com/stancld/rossum-mcp/pull/50)
+- Added `get_relation` and `list_relations` methods for `rossum_api.models.relation.Relation` objects. [#50](https://github.com/stancld/rossum-mcp/pull/50)
+- Added `get_document_relation` and `list_document_relations` methods for `rossum_api.models.document)relation.DcoumentRelation` objects. [#51](https://github.com/stancld/rossum-mcp/pull/51)